Started rewriting memory map to use preallocated space.
This approach is used in Linux kernel and I think it is easier to
grasp and maintain later on, as well as it practically removes need
to use allocator for other things than getting next free frame,
which should greately simplify workings of Booter.

The whole implementation will probably be just moved to GANDALF
reference implementation later on, which should make Booter
redundant as soon as GANDALF is at usable stage. Hence, this
reimplementation of Booter central algorithms will result in
(hopefully) last version of Booter - at least if GANDALF project
will became true.

Anyway, as for now, everything is broken and nothing compiles; I
do not want to start this in clean way as I did before, because
some parts will still be useful - but not really much of them,
as this seems to be yet another redesign of implementation of
this last stage bootloader.

Prototype kernel initialization routine soon to follow.